+++ app/scripts/comp/launcher.js
@@ -71,6 +71,12 @@ if (window.process && window.process.versions && window.process.versions.electro
         },
         cancelRestart: function() {
             this.restartPending = false;
+        },
+        getClipboardText: function() {
+            return this.req('clipboard').readText();
+        },
+        clearClipboardText: function() {
+            return this.req('clipboard').clear();
         }
     };
     window.launcherOpen = function(path) {