+++ app/scripts/comp/launcher.js
@@ -77,6 +77,12 @@ if (window.process && window.process.versions && window.process.versions.electro
         },
         clearClipboardText: function() {
             return this.req('clipboard').clear();
+        },
+        minimizeApp: function() {
+            this.remReq('app').minimizeApp();
+        },
+        canMinimize: function() {
+            return process.platform === 'win32';
         }
     };
     window.launcherOpen = function(path) {