+++ app/scripts/models/file-model.js
@@ -255,6 +255,7 @@ var FileModel = Backbone.Model.extend({
     },
 
     getData: function(cb) {
+        this.db.cutHistory();
         this.data = this.db.save(cb);
         return this.data;
     },