+++ app/scripts/presenters/entry-presenter.js
@@ -28,6 +28,7 @@ EntryPresenter.prototype = {
     get created() { return this.entry ? Format.dtStr(this.entry.created) : undefined; },
     get updated() { return this.entry ? Format.dtStr(this.entry.updated) : undefined; },
     get expired() { return this.entry ? this.entry.expired : false; },
+    get tags() { return this.entry ? this.entry.tags : false; },
     get description() {
         if (!this.entry) {
             return '[Group]';